Add initial EVM configuration
Added the github.com/evmos/ethermint evm and feemarket modules,
the github.com/Canto-Network/Canto erc20 module, lots of Canto and Evmos
cmd configuration, and custom combination of Evmos + Cosmos-SDK
initialization to get a minimal, working EVM on Althea chain.

Notably this implementation lacks the FeeGrant, Recovery, Epochs,
Inflation, Fees, and GovShuttle modules that Canto uses. As well the tps
counter has not been added.

Temporarily a fork of the Canto repo is on the althea-net GitHub to fix
an issue with their versioning mismatch (they are currently on v5.0.0
yet go is only aware of a v2). This fork will be dropped once they fix
their versions.