Update Native Token denom, integration test metadata
The Althea Chain testnet launched with a native staking token called
althea with 18 decimal precision, but due to genesis validation errors
the base denom is called ualthea (micro) instead of aalthea (atto).

This commit changes all references to ualtg or ualthea to use a new
aalthea denom, and also updates footoken to adhere to these principles
as well:
  Base: footoken -> ufootoken
  Display: mfootoken -> footoken
Additionally, aalthea has an alias called althea-wei and the althea
token has an additional denom unit called nalthea with an alias of
althea-gwei.