Adjust App setup to fix EVM bug + mirror Canto
With the 1.0+ updates to Web30 and Clarity the LegacyTx submission
failed on Althea L1 while succeeding on other EVMs. By tracing the app
initialization it was discovered that the EVM module lacked minter and
burner permissions. Beyond that, the updated organization has more
coherent structure and is easier to debug moving forward.