Implement Legacy Amino interfaces for x/microtx Msgs
For AltheaJS (EIP-712 message submissions) to work, LegacyAmino
en- and de-coding support is needed for all custom Msg types. This
commit resolves a confusing error on message submission:

    "raw_log":"expected *legacytx.LegacyMsg when using amino JSON: panic"

The error is produced in the x/auth module's legacy directory, in the
StdSignBytes function.
With the LegacyMsg interface implemented this error disappears and
signature verification proceeds as expected.