Add placeholder contract so other projects can import Althea-L1
contracts/liquid_account.go embeds a compiled contract definition, but
we did not add a precompiled definition to avoid the risk of an
out-of-date compilation being used in mainnet. Unfortunately this means
that other Go projects cannot import Althea-L1 because compilation
fails. This version of LiquidInfrastructureNFT is invalid but it will
prevent compilation failures.