Default to BaseAccount but replace with EthAccount
Keplr (and likely other wallets) expect to see Cosmos account types
(e.g. vesting or BaseAccount types) for Cosmos chains but error when the
sequence query returns an EthAccount type. To avoid this issue we
default to Cosmos BaseAccounts on creation but update the account type
when setting the PubKey if it is an Ethermint account. Vesting accounts
are left unchanged.