Add gentx signature validation to validate-genesis
This patch adds signature validation to the validate-genesis command,
speciifcally to the custom validations we already have for Althea chain.

Other validations on the gentx are performed in the collect gentx's
command, which ensures that the accounts exist and have the required
balances as part of placing the genesis transactions into the genesis.

But the collect gentx command does not actually validate the signatures
of those transactions. Leaving you vulnerable to an incorrect chainid
or otherwise bad signature.

While it's not ideal to have the signature verification in a different
command than collect-gentx this was the most easily accessible way to
add the feature. Ideally we would also do some balance verification
here. We do at least indirectly verify that the account exists, since
the sequence number lookup would fail if it does not.