Fix account generation for ethermint keys, add ante tests
This change requires an update to althea-net/ethermint so that the EVM
only creates the correct kind of accounts. When we set BaseAccount as
the default account type, it broke contract deployment because the
EthAccount is what makes code (especially the code hash) deployable at
an address.

Because ethermint was updated it was necessary to update the canto fork
as well.

The new paradigm initializes the evmKeeper with a special prototypical
account creation function to create a new EthAccount with an address
set. This function is used in the evm module and also throughout the
ante handlers everywhere a new account is created to ensure Ethermint
keys and contract accounts are all initialized correctly.

To ensure these changes work, ante handler tests have been copied and
enhanced from the ethermint project. I copied the general ante handler
tests but also added ones for the new EthSetPubkey and SetAccountType
decorators. The old antehandler is also checked to ensure the changes
fix the issue.