Add onboarding module tests
These tests are largely a copy of the Canto onboarding tests but without
any of the swap logic and many minor alterations to make them work with
the Althea-L1 configuration. Notably if the denom is not "aalthea" the tests
fail, and some of the setup tasks needed to be altered as well.