Rework go CI tests
The binary and go testing depends on the solidity folder contents
because we embed contracts for use by the x/erc20 module. To fix the go
CI tasks they must all install node.js and compile the solidity folder
before formatting the output for use by the go files in contracts. After
this the build, lint, and test tasks can all run.

Because they are no longer dependent on the build step, test and lint
are now completely parallel with build.