Fix: unknown method test
This test required no error for calling an endpoint that didn't exist.
Behavior wise this wasn't ever correct but previously the failure
occured at a different location in the call. Now it correctly occurs
when we try to call a method that obviously doesn't exist.

Contriburing to this is that calling "" is implicitly the constructor.
So the error was different than it normally would be for a unknown
method.