Check in compiled contracts required for build
This is unfortunately required for a reliable build experience. We've
done this before, gone back on this decision due to the lack of
transparency when checking in any binary output and then come back here
out of pragmatisim when it comes to the difficulty of actually executing
builds and ensuring those making their own builds don't have issues.

This contract bytecode is fully reproducbile by anyone else building
it, and explorer verification should also cross verify.