Override the sdk group Tx command submit-proposal
The sdk's group module CLI assumes that when a user is submitting a
proposal they have not and will never want to use the --from command.
Unfortunately that is not the case, so we invert the dependency
injection used when adding the Tx commands to the root cobra.Command to
override one of the module commands. This is preferable to forking the
entire SDK repo or duplicating the contents of the group module and
forever maintaining the whole module.

Unfortunately this means we will not get any changes to the
submit-proposal command in the future if any happen on upgrade, so we
will need to monitor for changes to this command.