Enhance the vote command, fix --exec flag
The vote command expects strings like VOTE_OPTION_YES from the user,
this commit adds a simple "yes" option that can be specified too. This
works for "no" "abstain" and "no-with-veto" as well.

The --exec flag help text was completely wrong, telling users to provide
"1" as a string when it only supported "try". Now both "1" and "try" are
supported.