Add backwards compatible codec info
The migration from canto to 'althea' erc20 module broke compatibility
for proposals that have passed on production althea l1 breaking all queries
for proposals that would include that data since it could not longer be
deserialized. This patch adds backwards compatible deserialization that
we will have to carry in the main binary until we perform a store
migration to the new types for the old proposals.

It's not really needed to migrate them ever, but would be required to
remove this code.